Add files via upload
There is a need to distinguish the mysql scripts which can execute with the results as we need. The mysql scripts are written in the excel.So the first step is to extract the scripts from it.Later,make a bit of modifications for the scripts to make sure it can be execute correctly via the mysql executing in the linux.Last but not least,create a result collection is also needed. So,saving the right and not right scripts in the different txts in different directories.